// src/enums.ts
/**
 * Состояния жизненного цикла страницы.
 */
export enum LifecycleState {
  /** Активное состояние, если страница видна и имеет фокус ввода. **/
  ACTIVE = 'active',
  /** Пассивное состояние, если страница видна и не имеет фокуса ввода. **/
  PASSIVE = 'passive',
  /** Скрытое состояние, если страница не видна и не была заморожена. **/
  HIDDEN = 'hidden',
  /** Замороженное состояние. **/
  FROZEN = 'frozen',
  /** Закрытое состояние, если страница начала выгружаться и очищаться из памяти браузером. **/
  TERMINATED = 'terminated',
  /** Неизвестное состояние, fallback. **/
  UNKNOWN = 'unknown',
}